remove git_revison.c, move git_revision.h to core Library
