+++ Examples/minimum_user/src/src_user/c2a_main.c
@@ -1,4 +1,4 @@
-#include <src_core/git_revision.h>
+#include <src_core/Library/git_revision.h>
 #include <src_core/Library/print.h>
 #include <src_core/c2a_core_main.h>
 #include <src_core/System/TimeManager/time_manager.h>
