+++ Library/git_revision.h
